+++ src/tests/solidityTests/testLogs/test.logs.sol
@@ -1,0 +1,9 @@
+pragma solidity ^0.4.4;
+
+import "wafr/Test.sol";
+
+contract TestLogs is Test {
+  function test_logs() {
+    log_uint(uint(49999), "some message");
+  }
+}